requirements.txt
pydantic>=2.0.0
crawl4ai
streamlit
pandas
openai>=1.0.0
folium
streamlit-folium
python-dotenv
fastapi>=0.104.0
uvicorn>=0.24.0
apscheduler>=3.10.0
playwright>=1.39.0
asyncio
prefect>=2.0.0
